# spec/api-spec.yml
openapi: 3.0.0
info:
  title: "GoMoney Football-League Manager Service"
  description: "This service provides endpoints for all `football-league manager` related interactions"
  version: 1.0.0
  contact:
    name: "AKEJU FATAI"
    email: "fatai@mail.com"
...
servers:
  - description: Staging Ingress
    url: https://api.gomoney
...
security:
  - bearerToken: [ ]
paths:
...
  /fixtures:
    post:
      summary: "Create Fixture"
      description: "Create Fixture"
      operationId: createFixture
      x-visibility: "public"
      tags:
        - "fixtures"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFixtureDto"
      responses:
        "201":
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "409":
          $ref: "#/components/responses/ConflictError"
        "default":
          $ref: "#/components/responses/ServerError"

    get:
      summary: "Get Fixtures"
      description: "Get Fixtures (Maybe Paginated Records - depends on the size of records)"
      operationId: getFixtures
      x-visibility: "public"
      parameters:
        - $ref: "#/components/parameters/fixtureStatusQuery"
        - $ref: "#/components/parameters/searchValueQuery"
        - $ref: "#/components/parameters/startDateQuery"
        - $ref: "#/components/parameters/endDateQuery"
        - $ref: "#/components/parameters/pageQuery"
        - $ref: "#/components/parameters/limitQuery"
        # - $ref: "#/components/parameters/sortQuery"
      tags:
        - "fixtures"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixtureListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "default":
          $ref: "#/components/responses/ServerError"
  /fixtures/{fixtureId}:
    parameters:
      - $ref: "#/components/parameters/fixtureIdParam"

    patch:
      summary: "Update Fixture"
      description: "Update Fixture"
      operationId: updateFixture
      x-visibility: "public"
      tags:
        - "fixtures"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFixtureDto"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ConflictError"
        "default":
          $ref: "#/components/responses/ServerError"

    get:
      summary: "Get Fixture"
      description: "Get Fixture"
      operationId: getFixture
      x-visibility: "public"
      tags:
        - "fixtures"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "default":
          $ref: "#/components/responses/ServerError"

    delete:
      summary: "Remove Fixture"
      description: "Remove Fixture"
      operationId: removeFixture
      x-visibility: "public"
      tags:
        - "fixtures"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDto"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "default":
          $ref: "#/components/responses/ServerError"
...
  /fixtures/by-status/{fixtureStatus}:
    parameters:
      - $ref: "#/components/parameters/fixtureStatusParam"
      - $ref: "#/components/parameters/pageQuery"
      - $ref: "#/components/parameters/limitQuery"

    get:
      summary: "Get Fixtures By Status"
      description: "Get Fixtures By Status"
      operationId: getFixturesByStatus
      x-visibility: "public"
      tags:
        - "fixtures"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixtureListResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "default":
          $ref: "#/components/responses/ServerError"
...
components:
  schemas:
...
    CreateFixtureDto:
      type: "object"
      description: "Create Fixture request data"
      additionalProperties: false
      required:
        - "homeTeamId"
        - "awayTeamId"
        - "commencesAt"
      properties:
        homeTeamId:
          type: "string"
          format: "object-id"
          description: "ID of Home Team"
        awayTeamId:
          type: "string"
          format: "object-id"
          description: "ID of Away Team"
        commencesAt:
          type: "string"
          format: "date-time"
          description: "Date & Time that match starts"
    UpdateFixtureDto:
      type: "object"
      description: "Update Fixture request data"
      additionalProperties: false
      minProperties: 1
      properties:
        homeTeamId:
          type: "string"
          format: "object-id"
          description: "ID of Home Team"
        awayTeamId:
          type: "string"
          format: "object-id"
          description: "ID of Away Team"
        commencesAt:
          type: "string"
          format: "date-time"
          description: "Date & Time that match starts"
        matchResult:
          $ref: "#/components/schemas/MatchResult"

    # RESPONSE OBJECT DEFINITIONS
...
    Fixture:
      type: "object"
      description: "Fixture object"
      required:
        - "id"
        - "homeTeam"
        - "awayTeam"
        - "commencesAt"
        - "status"
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Fixture unique identifier"
        homeTeam:
          type: "string"
          format: "uuid"
          description: "Home Team name"
        awayTeam:
          type: "string"
          format: "uuid"
          description: "Away Team nam"
        commencesAt:
          type: "string"
          format: "date-time"
          description: "Date & Time that match starts"
        status:
          $ref: "#/components/schemas/FixtureStatus"
        matchResult:
          $ref: "#/components/schemas/MatchResult"
        createdBy:
          type: "string"
          format: "uuid"
        updatedBy:
          type: "string"
          format: "uuid"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
    FixtureListResponse:
      # TO BE UPDATED TO THE PAGINATED RESPONSE FORMAT
      type: "object"
      required:
        - records
      properties:
        records:
          type: "array"
          items:
            $ref: '#/components/schemas/Fixture'
    MatchResult:
      type: "object"
      description: "Match-Result object"
      required:
        - "homeTeamScore"
        - "awayTeamScore"
      properties:
        homeTeamScore:
          type: "integer"
          minimum: 0
          description: "Score of Home Team"
        awayTeamScore:
          type: "integer"
          minimum: 0
          description: "Score of Away Team"
...
    StatusDto:
      description: Status data, check additional message field.
      type: object
      required:
        - success
      properties:
        success:
          description: >-
            Status can be successful or failed, a value of true indicates
            success.
          type: boolean
          example: true
        message:
          type: string
          description: additional message describing status.
        data:
          type: object
          additionalProperties:
            type: object
            properties: {}
          description: Data associated with the status, this will not always be present
        error:
          $ref: '#/components/schemas/ErrorMessageDto'
      x-common-model: StatusDto
    ErrorMessageDto:
      description: >-
        Default error object for services. This gives consistent error object
        that all services may use.
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: UNKNOWN_ERROR
        message:
          type: string
          description: Descriptive error message
          example: Request could not be completed due to an error
        data:
          type: object
          description: Additional data for this error message.
          additionalProperties: true
          properties: {}
      x-common-model: ErrorMessageDto

    # ENUM DEFINITIONS
    FixtureStatus:
      type: "string"
      description: "Possible States or Statuses of a Fixture"
      enum: &FixtureStatus
        - PENDING
        - COMPLETED
  parameters:
...
    fixtureIdParam:
      in: "path"
      name: "fixtureId"
      schema:
        type: "string"
        format: "object-id"
      required: true
      description: "Fixture identifier in path"
    fixtureStatusParam:
      in: "path"
      name: "fixtureStatus"
      schema:
        $ref: "#/components/schemas/FixtureStatus"
      required: true
      description: "Fixture status in path"
    fixtureStatusQuery:
      in: "query"
      name: "fixtureStatus"
      schema:
        $ref: "#/components/schemas/FixtureStatus"
      required: false
      description: "Fixture status in query"
    searchValueQuery:
      in: "query"
      name: "searchValue"
      schema:
        type: "string"
      required: false
      description: "Search value to filter records in query"
    startDateQuery:
      in: "query"
      name: "startDate"
      schema:
        type: "string"
        format: "date"
      required: false
      description: "Start date to filter records in query"
    endDateQuery:
      in: "query"
      name: "endDate"
      schema:
        type: "string"
        format: "date"
      required: false
      description: "End date to filter records in query"
...
    pageQuery:
      in: "query"
      name: page
      schema:
        type: "integer"
        format: "int32"
        minimum: 1
        description: "Positive integer"
      required: false
      description: Page to be gotten based on size parameters
    limitQuery:
      in: "query"
      name: limit
      schema:
        type: "integer"
        format: "int32"
        minimum: 1
        description: "Positive integer"
      required: false
      description: Number of items to be gotten per page
    sortQuery:
      in: "query"
      name: sort
      schema:
        type: "string"
        pattern: '^([a-zA-Z]{2,}(:asc|:desc)?)+(?:(,([a-zA-Z]{2,}(:asc|:desc)?))*)$'
      required: false
      allowReserved: true
      description: Sort string used to sort various fields in `asc` or `desc` order. E.g. `name:asc,createdAt:desc`
  responses:
...
    InvalidRequest:
      description: Request could not be validated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessageDto"
          example:
            code: 'BAD_REQUEST'
            message: 'Request could not be validated.'
    Unauthenticated:
      description: Client is not authenticated to make request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessageDto"
          example:
            code: 'UNAUTHENTICATED'
            message: 'Either client security header is missing or it is not valid.'
    AccessDenied:
      description: Expected permission is not available.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessageDto"
          example:
            code: 'ACCESS_DENIED'
            message: 'Expected permission or scope is not present.'
    NotFound:
      description: Entity was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessageDto"
          example:
            code: 'NOT_FOUND'
            message: 'Information could not be found'
    ConflictError:
      description: Entity already exist while user tries to create it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessageDto"
          example:
            code: 'ACCESS_DENIED'
            message: 'Entity already exist!'
...
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorMessageDto"
          example:
            code: 'INTERNAL_ERROR'
            message: 'An internal server error occurred.'
  securitySchemes:
    bearerToken:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "Authentication for user-facing endpoints. Uses bearer token in API prefix."